pyproject.toml
[project]
name = "ox-ai"
version = "0.1.0"
description = "A lightweight framework for managing AI research experiments with Claude Code"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Tom Firth", email = "thomas.d.firth@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "duckdb>=1.0",
]

[project.optional-dependencies]
wandb = ["wandb>=0.16"]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
ox = "ox.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ox"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
